Refactor pkgcommand script
