+++ dev/pyintel/__init__.py
@@ -1,2 +1,3 @@
 from .uni import generate_unimathsymbols_intel
 from .ctanpkg import CtanPkg
+from .pkgcommand import CwlIntel
